# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/object": {
            "get": {
                "description": "get an object from the store tha can be released later",
                "summary": "get object",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    }
                }
            },
            "post": {
                "description": "create  an object into the store",
                "consumes": [
                    "application/json"
                ],
                "summary": "create a new object in the store",
                "parameters": [
                    {
                        "description": "object to create",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        },
        "/object/{obj}": {
            "put": {
                "description": "return an object into the store",
                "summary": "return the object",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "object to return",
                        "name": "obj",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "http.CreateReq": {
            "type": "object",
            "properties": {
                "obj": {
                    "type": "integer"
                }
            }
        }
    }
}